Optimize debug_executionWitness by pre-serializing RPC format at storage time
Move RpcExecutionWitness to ethrex-common and convert ExecutionWitness to RPC
format when storing witnesses. This eliminates the expensive encode_subtrie()
traversal on every read by pre-computing it once at write time.

Key changes:
- store_witness() now converts ExecutionWitness → RpcExecutionWitness before storing
- Added get_witness_json_bytes() to return raw JSON bytes for zero-copy RPC response
- RPC handler now parses cached bytes directly to serde_json::Value

Expected improvement: ~70-110ms reduction on cached witness reads by eliminating
encode_subtrie() DFS traversal and avoiding re-serialization.
