+++ crates/networking/rpc/clients/eth/mod.rs
@@ -1,7 +1,6 @@
 use std::collections::BTreeMap;
 
 use crate::{
-    debug::execution_witness::RpcExecutionWitness,
     eth::client::EthConfigResponse,
     mempool::MempoolContent,
     types::{
@@ -16,7 +15,10 @@ use bytes::Bytes;
 use errors::{EthClientError, RpcRequestError};
 use ethrex_common::{
     Address, H256, U256,
-    types::{AuthorizationTupleEntry, BlobsBundle, Block, GenericTransaction, TxKind},
+    types::{
+        AuthorizationTupleEntry, BlobsBundle, Block, GenericTransaction, TxKind,
+        block_execution_witness::RpcExecutionWitness,
+    },
     utils::decode_hex,
 };
 use ethrex_rlp::decode::RLPDecode;
